requirements.txt
discord.py>=0.16.8,<0.17
